Fix custom shift date interval causing wrong data line
