+++ src/client/app/translations/data.ts
@@ -48,7 +48,7 @@ const LocaleTranslationData = {
 		"clipboard.copied": "Copied To Clipboard",
 		"clipboard.not.copied": "Failed to Copy To Clipboard",
 		"close": "Close",
-		"compare": "Compare",
+		"compare": "Compare bar",
 		"compare.raw": "Cannot create comparison graph on raw units such as temperature",
 		"confirm.action": "Confirm Action",
 		"contact.us": "Contact us",
@@ -561,7 +561,7 @@ const LocaleTranslationData = {
 		"clipboard.copied": "Copied To Clipboard\u{26A1}",
 		"clipboard.not.copied": "Failed to Copy To Clipboard\u{26A1}",
 		"close": "Close\u{26A1}",
-		"compare": "Comparer",
+		"compare": "Compare bar\u{26A1}",
 		"compare.raw": "Cannot create comparison graph on raw units such as temperature\u{26A1}",
 		"confirm.action": "Confirm Action\u{26A1}",
 		"contact.us": "Contactez nous",
@@ -1073,7 +1073,7 @@ const LocaleTranslationData = {
 		"clipboard.copied": "Copiado al portapapeles",
 		"clipboard.not.copied": "Error al copiar al portapapeles",
 		"close": "Cerrar",
-		"compare": "Comparar",
+		"compare": "Compare bar\u{26A1}",
 		"compare.raw": "No se puede crear un gráfico de comparación con unidades crudas como temperatura",
 		"confirm.action": "Confirmar acción",
 		"contact.us": "Contáctenos",
